functions/requirements.txt
flask
flask-restplus
flask-restx
firebase-admin
firebase-functions
python-dotenv
requests
werkzeug
